+++ CHANGELOG.md
@@ -51,6 +51,7 @@ This changelog starts with version 2.0.0. Basically all parts of the system, inc
   - fixed inconsistent history meta writing
   - fixed handling of existing flags
 - `proGeneeric`: fixed inconsistent history meta writing
+- `sphinx-doc`: removed documentation of data/flags parameters from automatic sphinx doc
 
 ## [2.0.1](https://git.ufz.de/rdm-software/saqc/-/tags/v2.0.1) - 2021-12-20
 [List of commits](https://git.ufz.de/rdm-software/saqc/-/compare/v2.0.0...v2.0.1)